refactor: 將DTO提出去放
