+++ src/service/user/autoLoginToGetTenantToken.ts
@@ -5,6 +5,7 @@ import { loginByTenantId, loginSystemAdmin } from '../../library/thingsboardConn
 import getTenantList from './helpers/getTenantList';
 import { SystemAdminProfileProps } from '../../interface/user';
 import isTenantEntityNotFound from './helpers/isTenantEntityNotFound';
+import AutoLoginToGetTenantTokenDTO from '../../interface/serviceResponse/autoLoginToGetTenantTokenDTO';
 
 // eslint-disable-next-line no-unused-vars
 const loggers = new WinstonLogger({ type: 'User service' });
@@ -14,42 +15,6 @@ const defaultAdminProfile: SystemAdminProfileProps = {
     username: TB_USER.systemAdminEmail,
     password: TB_USER.systemAdminPassword,
 };
-
-interface GetTenantTokenRes {
-    status: number;
-    systemAdminEmail: string;
-    tenantAdminName: string;
-    tenantEmail: string;
-    token: string;
-    refreshToken: string;
-    errorMessage: any;
-}
-
-class AutoLoginToGetTenantTokenDTO implements GetTenantTokenRes {
-    status: number;
-
-    systemAdminEmail: string;
-
-    tenantAdminName: string;
-
-    tenantEmail: string;
-
-    token: string;
-
-    refreshToken: string;
-
-    errorMessage: any;
-
-    constructor(data: any) {
-        this.status = data.status;
-        this.systemAdminEmail = data.systemAdminEmail;
-        this.tenantAdminName = data.tenantAdminName;
-        this.tenantEmail = data.tenantEmail;
-        this.token = data.token;
-        this.refreshToken = data.refreshToken;
-        this.errorMessage = data.errorMessage;
-    }
-}
 
 export default async function autoLoginToGetTenantToken() {
     // 1. 取得Admin token
